Resolve "Initialize weights on device side"

+++ HugeCTR/include/utils.cuh
@@ -71,7 +71,9 @@ struct TypeConvertFunc<float, long long> {
 
 template <>
 struct TypeConvertFunc<float, unsigned int> {
-  static __forceinline__ __device__ float convert(unsigned int val) { return static_cast<float>(val); }
+  static __forceinline__ __device__ float convert(unsigned int val) {
+    return static_cast<float>(val);
+  }
 };
 
 template <typename T>
@@ -105,28 +107,28 @@ __inline__ __device__ T blockReduceSum(T val) {
 }
 
 template <typename T>
-__global__ void initialize_array(T* array, int num_elements, T value) {
-  const int tid_base = blockIdx.x * blockDim.x + threadIdx.x;
-  const int num_threads = blockDim.x * gridDim.x;
-  for (int tid = tid_base; tid < num_elements; tid += num_threads) {
+__global__ void initialize_array(T* array, size_t num_elements, T value) {
+  const size_t tid_base = blockIdx.x * blockDim.x + threadIdx.x;
+  const size_t num_threads = blockDim.x * gridDim.x;
+  for (size_t tid = tid_base; tid < num_elements; tid += num_threads) {
     array[tid] = value;
   }
 }
 
 template <typename TIN, typename TOUT, typename Lambda>
-__global__ void transform_array(const TIN* in, TOUT* out, int num_elements, Lambda op) {
-  const int tid_base = blockIdx.x * blockDim.x + threadIdx.x;
-  const int num_threads = blockDim.x * gridDim.x;
-  for (int tid = tid_base; tid < num_elements; tid += num_threads) {
+__global__ void transform_array(const TIN* in, TOUT* out, size_t num_elements, Lambda op) {
+  const size_t tid_base = blockIdx.x * blockDim.x + threadIdx.x;
+  const size_t num_threads = blockDim.x * gridDim.x;
+  for (size_t tid = tid_base; tid < num_elements; tid += num_threads) {
     out[tid] = op(in[tid]);
   }
 }
 
 template <typename TIN, typename TOUT>
-__global__ void convert_array(TOUT* out, const TIN* in, int num_elements) {
-  const int tid_base = blockIdx.x * blockDim.x + threadIdx.x;
-  const int num_threads = blockDim.x * gridDim.x;
-  for (int tid = tid_base; tid < num_elements; tid += num_threads) {
+__global__ void convert_array(TOUT* out, const TIN* in, size_t num_elements) {
+  const size_t tid_base = blockIdx.x * blockDim.x + threadIdx.x;
+  const size_t num_threads = blockDim.x * gridDim.x;
+  for (size_t tid = tid_base; tid < num_elements; tid += num_threads) {
     out[tid] = TypeConvertFunc<TOUT, TIN>::convert(__ldg(in + tid));
   }
 }
